requirements.txt
daemonize
jinja2
pyyaml
requests
schedule
slacker
termcolor
